Color according to n-log2(log(r^2)).  Still steppy.  Fixed point version missing.

+++ lib/mandy.h
@@ -19,7 +19,7 @@
 #include <config.h>
 #include "Fixed.h"
 
-typedef int count_t;
+typedef double count_t;
 
 void fatal(int errno_value, const char *fmt, ...);
